Pre-release clean-up - 39 Plan & Pricing - 15

+++ utils/saveutils.cpp
@@ -369,9 +369,9 @@ QByteArray projectTheme(const QString& projectDir)
     return Internal::property(projectDir, ProjectTheme).value<QByteArray>();
 }
 
-quint32 userPlan(const QString& userDir)
+qint64 userPlan(const QString& userDir)
 {
-    return Internal::property(userDir, UserPlan).value<quint32>();
+    return Internal::property(userDir, UserPlan).value<qint64>();
 }
 
 QString userEmail(const QString& userDir)
