+++ utils/saveutils.h
@@ -72,7 +72,7 @@ QDateTime projectCreationDate(const QString& projectDir);
 QDateTime projectModificationDate(const QString& projectDir);
 QByteArray projectTheme(const QString& projectDir);
 
-quint32 userPlan(const QString& userDir);
+qint64 userPlan(const QString& userDir);
 QString userEmail(const QString& userDir);
 QByteArray userPassword(const QString& userDir);
